Create python-publish.yml
This introduces the script to publish a PyPI package whenever a new release is made on the Github repo. When a new release is made (ie. not when a commit a tagged) the PyPI key is used by the action to push the new package, pre-releases should work as well. 

Signed-off-by: Eric Kerfoot <17726042+ericspod@users.noreply.github.com>